move RAILDIR from core.utils to core.common_params and change all the import path
